Change key datatype for readings
This will allow for many more readings to be stored. Probably enough
key space for storing interval readings directly from sensors

+++ backend/src/api/graphql/Reading.ts
@@ -10,7 +10,7 @@ export const Reading = objectType({
     name: 'Reading',
     sourceType: 'ReadingModel',
     definition(t) {
-        t.nonNull.int('id'),
+        t.nonNull.bigint('id'),
         t.nonNull.float('value'),
         t.nonNull.field('unit', { type: ReadingUnitEnum }),
         t.nonNull.field('location', {
